[clean]ソースの整理
コメントの追加やコメント、無駄な改行の削除

+++ source/Score.h
@@ -11,8 +11,8 @@ private:
 	int score;			  //スコア
 	int totalShot;		  //弾を撃った数
 	int totalHit;		  //弾を当てた数
-	float oneShotRatio;	  //弾一発の割合
-	float hitMissRatio;	  //弾を外した割合
+	double oneShotRatio;  //弾一発の割合
+	double hitMissRatio;  //弾を外した割合
 	int scoreX;			  //スコアがずれていく用
 
 	//snprintf用
